Separate serialization logic from class
